pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = ""
authors = ["Group_6"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-environ = "^0.11.2"
pydrive = "^1.3.1"
psycopg2 = "^2.9.9"
aiogoogle = "^5.7.0"
redis = {extras = ["hiredis"], version = "^5.0.4"}
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
